Consolidated duplicate code in weapon list fragments
WeaponListFragment was unused, so it was made into the superclass of the
other weapon fragments, and the duplicate code was moved there. While
using delegation would be better, this is easier to implement and may
make it simpler to add new changes in the future.
