+++ app/src/main/java/com/daviancorp/android/ui/list/WeaponListFragment.java
@@ -1,84 +1,149 @@
 package com.daviancorp.android.ui.list;
 
-import android.support.v4.content.AsyncTaskLoader;
 import android.content.Context;
-import android.content.Intent;
-import android.database.Cursor;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v4.app.ListFragment;
-import android.support.v4.app.LoaderManager.LoaderCallbacks;
+import android.support.v4.app.LoaderManager;
+import android.support.v4.content.AsyncTaskLoader;
 import android.support.v4.content.Loader;
-import android.support.v4.widget.CursorAdapter;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.ListView;
+import android.view.ViewGroup;
 
-import com.daviancorp.android.data.classes.Weapon;
 import com.daviancorp.android.data.database.DataManager;
-import com.daviancorp.android.loader.WeaponListCursorLoader;
 import com.daviancorp.android.mh4udatabase.R;
-import com.daviancorp.android.ui.detail.WeaponDetailActivity;
+import com.daviancorp.android.ui.adapter.WeaponExpandableListGeneralAdapter;
+import com.daviancorp.android.ui.general.WeaponListEntry;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
+import java.util.List;
 
-public abstract class WeaponListFragment extends ListFragment implements
-		LoaderCallbacks<ArrayList<Weapon>> {
-	protected static final String ARG_TYPE = "WEAPON_TYPE";
+/**
+ * The superclass for weapons list fragments of every weapon type.
+ */
+public abstract class WeaponListFragment extends Fragment implements
+        LoaderManager.LoaderCallbacks<ArrayList<WeaponListEntry>> {
 
-//	private static final String DIALOG_WISHLIST_DATA_ADD_MULTI = "wishlist_data_add_multi";
-//	private static final int REQUEST_ADD_MULTI = 0;
+    protected static final String ARG_TYPE = "WEAPON_TYPE";
+    private static final String GROUPS_KEY = "groups_key";
 
-	@Override
-	public void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-	}
+    private LinearLayoutManager mLayoutManager;
+    private Bundle savedState;
 
-	
-	@Override
-	public Loader<ArrayList<Weapon>> onCreateLoader(int id, Bundle args) {
-		// You only ever load the runs, so assume this is the case
-		String mType = null;
-		if (args != null) {
-			mType = args.getString(ARG_TYPE);
-		}
-		return new WeaponArrayLoader(getActivity().getApplicationContext(), mType);
-	}
+    // todo: refactor so that these variables don't have to be set to protected
+    protected WeaponExpandableListGeneralAdapter mAdapter;
+    protected RecyclerView mRecyclerView;
 
-	@Override
-	public void onLoaderReset(Loader<ArrayList<Weapon>> loader) {
-		// Stop using the cursor (via the adapter)
-		setListAdapter(null);
-	}
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
 
-	@Override
-	public abstract void onLoadFinished(Loader<ArrayList<Weapon>> loader, ArrayList<Weapon> weapons);
-	
-	@Override
-	public void onListItemClick(ListView l, View v, int position, long id) {
-		// The id argument will be the Weapon ID; CursorAdapter gives us this
-		// for free
-		Intent i = new Intent(getActivity(), WeaponDetailActivity.class);
-		i.putExtra(WeaponDetailActivity.EXTRA_WEAPON_ID, id);
-		startActivity(i);
-	}
+        setHasOptionsMenu(true);
 
-    static class WeaponArrayLoader extends AsyncTaskLoader<ArrayList<Weapon>> {
+        savedState = savedInstanceState;
+
+        // Initialize the loader to load the list of runs
+        getLoaderManager().initLoader(R.id.weapon_list_fragment, getArguments(), this).forceLoad();
+    }
+
+    /**
+     * Used by onCreateView to build an adapter that creates the actual views for all weapons.
+     * @return
+     */
+    protected abstract WeaponExpandableListGeneralAdapter createWeaponAdapter();
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View rootView = inflater.inflate(R.layout.fragment_weapon_tree_expandable, container, false);
+        mRecyclerView = (RecyclerView) rootView.findViewById(R.id.content_recyclerview);
+        mRecyclerView.setHasFixedSize(true);
+
+        mLayoutManager = new LinearLayoutManager(getActivity());
+        mRecyclerView.setLayoutManager(mLayoutManager);
+
+        mAdapter = createWeaponAdapter();
+        mRecyclerView.setAdapter(mAdapter);
+
+        // Restores old groups if we are returning to the fragment
+        if (savedInstanceState != null) {
+            List<Integer> groups = savedInstanceState.getIntegerArrayList(GROUPS_KEY);
+            mAdapter.restoreGroups(groups);
+        }
+
+        return rootView;
+    }
+
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+
+        // Restores old groups if we are returning to the fragment
+        if (savedInstanceState != null) {
+            List<Integer> groups = savedInstanceState.getIntegerArrayList(GROUPS_KEY);
+            mAdapter.restoreGroups(groups);
+        }
+    }
+
+    //
+    // Saves current state of groups
+    //
+    @Override
+    public void onSaveInstanceState(Bundle outState) {
+        outState.putIntegerArrayList(GROUPS_KEY, mAdapter.saveGroups());
+        super.onSaveInstanceState(outState);
+    }
+
+    @Override
+    public void onLoadFinished(Loader<ArrayList<WeaponListEntry>> loader,
+                               ArrayList<WeaponListEntry> weapons) {
+        mRecyclerView.setAdapter(mAdapter);
+        mAdapter.addAll(weapons);
+    }
+
+    @Override
+    public void onLoaderReset(Loader<ArrayList<WeaponListEntry>> loader) {
+        // Stop using the cursor (via the adapter)
+        mRecyclerView.setAdapter(null);
+        mAdapter.clear();
+    }
+
+    @Override
+    public Loader<ArrayList<WeaponListEntry>> onCreateLoader(int id, Bundle args) {
+        // You only ever load the runs, so assume this is the case
+        String mType = null;
+        if (args != null) {
+            mType = args.getString(ARG_TYPE);
+        }
+        return new WeaponArrayLoader(getActivity().getApplicationContext(), mType);
+    }
+
+    static class WeaponArrayLoader extends AsyncTaskLoader<ArrayList<WeaponListEntry>> {
         String mType;
         public WeaponArrayLoader(Context context, String type) {
             super(context);
             mType = type;
         }
 
-        public ArrayList<Weapon> loadInBackground() {
-            return DataManager.get(getContext()).queryWeaponTypeArray(mType);
+        public ArrayList<WeaponListEntry> loadInBackground() {
+            return DataManager.get(getContext()).queryWeaponTreeArray(mType);
         }
     }
 
+    @Override
+    public void onPause() {
+        savedState = new Bundle();
+        savedState.putIntegerArrayList(GROUPS_KEY, mAdapter.saveGroups());
+        super.onPause();
+    }
 
-	//protected abstract CursorAdapter getDetailAdapter();
-
-	//protected abstract Weapon getDetailWeapon(int position);
-
-	//protected abstract Fragment getThisFragment();
+    @Override
+    public void onResume() {
+        if (savedState != null) {
+            List<Integer> groups = savedState.getIntegerArrayList(GROUPS_KEY);
+            mAdapter.restoreGroups(groups);
+        }
+        super.onResume();
+    }
 }
